HWKBTM-416 Add a severity value to communication summary graph nodes and links
