+++ api/src/test/java/org/hawkular/btm/api/services/AbstractAnalyticsServiceTest.java
@@ -18,16 +18,12 @@ package org.hawkular.btm.api.services;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import org.hawkular.btm.api.model.analytics.CommunicationSummaryStatistics;
 import org.hawkular.btm.api.model.analytics.EndpointInfo;
@@ -265,84 +261,6 @@ public class AbstractAnalyticsServiceTest {
         assertFalse(result.get(0).metaURI());
 
         assertNotNull(result.get(0).getRegex());
-    }
-
-    @Test
-    public void testGetRootCommunicationSummaryNodes() {
-        Map<String, CommunicationSummaryStatistics> nodeMap = new HashMap<String, CommunicationSummaryStatistics>();
-
-        CommunicationSummaryStatistics css1 = new CommunicationSummaryStatistics();
-        css1.setId("css1");
-        nodeMap.put(css1.getId(), css1);
-
-        CommunicationSummaryStatistics.ConnectionStatistics con1 =
-                new CommunicationSummaryStatistics.ConnectionStatistics();
-        css1.getOutbound().put("css2", con1);
-
-        CommunicationSummaryStatistics css2 = new CommunicationSummaryStatistics();
-        css2.setId("css2");
-        nodeMap.put(css2.getId(), css2);
-
-        CommunicationSummaryStatistics css3 = new CommunicationSummaryStatistics();
-        css3.setId("css3");
-        nodeMap.put(css3.getId(), css3);
-
-        Collection<CommunicationSummaryStatistics> result =
-                AbstractAnalyticsService.getRootCommunicationSummaryNodes(nodeMap);
-
-        assertNotNull(result);
-        assertEquals(2, result.size());
-
-        for (CommunicationSummaryStatistics css : result) {
-            assertNotEquals("css2", css.getId());
-        }
-    }
-
-    @Test
-    public void testBuildCommunicationSummaryTree() {
-        List<CommunicationSummaryStatistics> nodes = new ArrayList<CommunicationSummaryStatistics>();
-
-        CommunicationSummaryStatistics css1 = new CommunicationSummaryStatistics();
-        css1.setId("css1");
-        nodes.add(css1);
-
-        CommunicationSummaryStatistics.ConnectionStatistics con1 =
-                new CommunicationSummaryStatistics.ConnectionStatistics();
-        css1.getOutbound().put("css2", con1);
-
-        CommunicationSummaryStatistics css2 = new CommunicationSummaryStatistics();
-        css2.setId("css2");
-        nodes.add(css2);
-
-        CommunicationSummaryStatistics css3 = new CommunicationSummaryStatistics();
-        css3.setId("css3");
-        nodes.add(css3);
-
-        Collection<CommunicationSummaryStatistics> result =
-                AbstractAnalyticsService.buildCommunicationSummaryTree(nodes);
-
-        assertNotNull(result);
-        assertEquals(2, result.size());
-
-        CommunicationSummaryStatistics css1a = null;
-        CommunicationSummaryStatistics css3a = null;
-
-        for (CommunicationSummaryStatistics css : result) {
-            if (css.getId().equals("css1")) {
-                css1a = css;
-            } else if (css.getId().equals("css3")) {
-                css3a = css;
-            } else {
-                fail("Unexpected id: " + css.getId());
-            }
-        }
-
-        assertEquals(1, css1a.getOutbound().size());
-        assertTrue(css1a.getOutbound().containsKey("css2"));
-        assertTrue(css3a.getOutbound().isEmpty());
-
-        assertNotNull(css1a.getOutbound().get("css2").getNode());
-        assertEquals("css2", css1a.getOutbound().get("css2").getNode().getId());
     }
 
     @Test
